Removed unused public info and organized files location
