+++ iOSAssignment/Models/ImageDetail.swift
@@ -1,12 +1,5 @@
-public struct ImageDetail: Codable {
-    public let url: String
-    public let width: UInt
-    public let height: UInt
-}
-
-public extension Array where Element == ImageDetail {
-
-    var smallestImage: ImageDetail? {
-        self.min(by: { $0.width < $1.width })
-    }
+struct ImageDetail: Codable {
+    let url: String
+    let width: UInt
+    let height: UInt
 }
