+++ iOSAssignment/Models/ProductPrice.swift
@@ -1,0 +1,11 @@
+//
+//  ProductPrice.swift
+//  iOSAssignment
+//
+//  Created by Jader Nunes on 25/10/21.
+//
+
+struct ProductPrice: Codable {
+    let currency: String
+    let amount: Int
+}
