+++ iOSAssignment/Models/ProductUnitPrice.swift
@@ -1,0 +1,12 @@
+//
+//  ProductUnitPrice.swift
+//  iOSAssignment
+//
+//  Created by Jader Nunes on 25/10/21.
+//
+
+struct ProductUnitPrice: Codable {
+    let price: ProductPrice?
+    let unit: String?
+    let unitAmount: String?
+}
